Started using workspace, added cryptography
Firstly, I changed the crate structure to be a workplace. This is way clearer. Now, there's a single binary file (main.rs) that controls everything, and two libraries (so far), one that archives and compresses, and a second that derives a key, encrypts, and saves metadata  (salts, nonces, MK's digest) to a TOML file.